Inline content utilities into www app
This change removes the separate @nx/content workspace package and moves the markdown navigation/content helpers directly into the www app. The app now reads frontmatter from local markdown files via apps/www/app/NX/lib/content.ts, updates server hooks and page imports to use the local module, and removes the obsolete package references and TypeScript aliases. This keeps the content logic co-located with the site and avoids the extra package dependency.
