+++ apps/www/tsconfig.json
@@ -21,8 +21,6 @@
     ],
     "paths": {
       "@/*": ["./*"],
-      "@nx/content": ["../../packages/content/src"],
-      "@nx/content/*": ["../../packages/content/src/*"],
       "@nx/design-system": ["../../packages/design-system/src"],
       "@nx/design-system/*": ["../../packages/design-system/src/*"],
       "@nx/uberedux": ["../../packages/uberedux/src"],
