--- packages/content/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "@nx/content",
-  "version": "3.0.3",
-  "description": "Shared content navigation utilities for markdown-driven NX apps.",
-  "private": true,
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
-  "exports": {
-    ".": {
-      "types": "./src/index.ts",
-      "import": "./src/index.ts",
-      "require": "./src/index.ts"
-    },
-    "./server": {
-      "types": "./src/server.ts",
-      "import": "./src/server.ts",
-      "require": "./src/server.ts"
-    }
-  },
-  "dependencies": {
-    "gray-matter": "^4.0.3"
-  },
-  "scripts": {
-    "test": "tsx --test tests/**/*.test.ts",
-    "type-check": "tsc --noEmit",
-    "lint": "echo 'No lint configured yet'"
-  },
-  "devDependencies": {
-    "@types/node": "^20",
-    "tsx": "^4.21.0",
-    "typescript": "^5"
-  }
-}
