+++ tsconfig.json
@@ -20,8 +20,6 @@
       "@nx/ui/*": ["packages/ui/*"],
       "@nx/design-system": ["packages/design-system/src"],
       "@nx/design-system/*": ["packages/design-system/src/*"],
-      "@nx/content": ["packages/content/src"],
-      "@nx/content/*": ["packages/content/src/*"],
       "@nx/shortcodes": ["packages/shortcodes/src"],
       "@nx/shortcodes/*": ["packages/shortcodes/src/*"],
       "@nx/uberedux": ["packages/uberedux/src"],
